schoolFactory and sample data
In  the final json file will be located with all school information.
Up to now only a sample json is added for developing and testing till the
final version is available.

 will hold the angular factory which is used to fetch the
json data and provides it to other controllers. This factory is used so
in the future only the data provider can be replaced with an database provider
if necessary without touching other files.